libcamera2 : new camera kernel/userland interface. Fix zoom bug, adapt settings to the sensor used by device (can be easely updated with new sensors). Fix the video recording.
The library (liboemcamera.so) used on the device can be the liboemcamera.so from the dream package (with a small binary patch) or the one from the legend package (with some binaries patches). It autodetects the library used by trying to link the external "mt9p012_process_start" subroutine. As the dream liboemcamera.so does not support the mt9p012 sensor, it will fail on this library. That's the way it's detected.