Revert "gralloc: increase buffer size for wvga (like froyo)"
This reverts commit aa01920dcb372839a43d86bfb8c31eb4e2d05108.

The commit in question is practically different than the message which went
with it: the buffer size was reduced and it was changed from what the size in
froyo is.

The value must match what is supported by our binary GL driver. Therefore we
will need to keep this as is for 3D to work correctly.