msm72xx_set_acoustic_table: Use Current device if device argument does not match any sound end points.
There are points in AudioHardware where in attempt to patch the audio_path to the current sound device, we erroneously pass (device=-1), which is an invalid argument leading to 'Uknown out_path'.  This patch should be coupled with a kerel patch to qdsp5/snd.c which hardens the check_device routine to check whether the device argument is in range on the lower end of the spectrum (<0).  The kernel patch prevents an all out audio crash occuring when msm_snd is ioctl-ed to SND_SET_DEVICE with argument (device=-1).